cleans up unit tests and fleshes out the integration test with all current sections used in this diff
